refactor: rename relay2 to relay
