+++ testing/demo/pkg/relay/main.go
@@ -2,127 +2,358 @@ package main
 
 import (
 	"context"
+	"crypto/ecdsa"
+	"encoding/hex"
 	"fmt"
 	"log"
+	"math/big"
+	"strings"
 	"time"
 
+	"cosmossdk.io/math"
 	proverclient "github.com/celestiaorg/celestia-zkevm-ibc-demo/provers/client"
 	"github.com/celestiaorg/celestia-zkevm-ibc-demo/testing/demo/pkg/utils"
+	sdktypes "github.com/cosmos/cosmos-sdk/types"
+	transfertypes "github.com/cosmos/ibc-go/v9/modules/apps/transfer/types"
 	channeltypesv2 "github.com/cosmos/ibc-go/v9/modules/core/04-channel/v2/types"
+	ibchostv2 "github.com/cosmos/ibc-go/v9/modules/core/24-host/v2"
+	ibctesting "github.com/cosmos/ibc-go/v9/testing"
+	"github.com/cosmos/solidity-ibc-eureka/abigen/ics20lib"
+	"github.com/cosmos/solidity-ibc-eureka/abigen/ics26router"
+	"github.com/cosmos/solidity-ibc-eureka/abigen/icscore"
+	"github.com/cosmos/solidity-ibc-eureka/abigen/sp1ics07tendermint"
+	"github.com/ethereum/go-ethereum/accounts/abi"
+	"github.com/ethereum/go-ethereum/accounts/abi/bind"
+	ethcommon "github.com/ethereum/go-ethereum/common"
+	ethtypes "github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/crypto"
+	"github.com/ethereum/go-ethereum/ethclient"
+	"github.com/srdtrk/solidity-ibc-eureka/e2e/v8/ethereum"
 	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials/insecure"
 )
 
+// TODO: fetch these from the `make setup` command output.
+// const (
+// 	ics26Router            = "0xe53275a1fca119e1c5eeb32e7a72e54835a63936"
+// 	icsCore                = "0x505f890889415cf041001f5190b7800266b0dddd"
+// 	ics07TMContractAddress = "0x25cdbd2bf399341f8fee22ecdb06682ac81fdc37"
+// )
+
 const (
-	// celestiaProverEndpoint is the endpoint of the Celestia prover.
-	celestiaProverEndpoint = "localhost:50051"
-	// channelID is the channel ID on SimApp.
-	// TODO: fetch this from the `make setup` command output.
-	channelID = "channel-0"
-	// ics07TMContractAddress is the contract address of the ICS07 light client on the EVM roll-up.
-	// TODO: fetch this from the `make setup` command output.
-	ics07TMContractAddress = "0x25cdbd2bf399341f8fee22ecdb06682ac81fdc37"
+	// sender is an address on SimApp that will send funds via the MsgTransfer.
+	sender = "cosmos1ltvzpwf3eg8e9s7wzleqdmw02lesrdex9jgt0q"
+	// receiver is an address on the EVM chain that will receive funds via the MsgTransfer.
+	receiver = "0x7f39c581f595b53c5cb19b5a6e5b8f3a0b1f2f6e"
+	// denom is the denomination of the token on SimApp.
+	denom = "stake"
+	// sourceChannel is hard-coded to the name used by the first channel.
+	sourceChannel = ibctesting.FirstChannelID
+	// sequence is hard-coded to the first sequence number.
+	sequence = 1
+	// ethereumRPC is the Reth RPC endpoint.
+	ethereumRPC = "http://localhost:8545/"
+	// ethereumAddress is an address on the EVM chain.
+	// ethereumAddress = "0xaF9053bB6c4346381C77C2FeD279B17ABAfCDf4d"
+	// ethPrivateKey is the private key for ethereumAddress.
+	ethPrivateKey = "0x82bfcfadbf1712f6550d8d2c00a39f05b33ec78939d0167be2a737d691f33a6a"
+	// cliendID is for the SP1 Tendermint light client on the EVM roll-up.
+	clientID = "07-tendermint-0"
 )
 
 func main() {
-	// Ask the Celestia prover for a state transition proof.
-	_, err := GetStateTransitionProof()
+	err := updateTendermintLightClient()
 	if err != nil {
 		log.Fatal(err)
 	}
 
-	packetResp, err := QueryPacketCommitments()
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	// Ask the Celestia prover for a state membership proof that the packet
-	// commitments are part of the state root at a particular block height.
-	_, err = GetMembershipProof(packetResp)
-	if err != nil {
-		log.Fatal(err)
-	}
-
+	// err = receivePacketOnEVM()
+	// if err != nil {
+	// 	log.Fatal(err)
+	// }
 }
 
-// QueryPacketCommitments queries the packet commitments on the SimApp.
-func QueryPacketCommitments() (*channeltypesv2.QueryPacketCommitmentsResponse, error) {
-	fmt.Printf("Querying packet commitments on SimApp...\n")
+// updateTendermintLightClient submits a MsgUpdateClient to the Tendermint light client on the EVM roll-up.
+func updateTendermintLightClient() error {
+	addresses, err := utils.ExtractDeployedContractAddresses()
+	if err != nil {
+		return err
+	}
+	fmt.Printf("addresses %v\n", addresses)
 
-	clientCtx, err := utils.SetupClientContext()
+	ethClient, err := ethclient.Dial(ethereumRPC)
+	if err != nil {
+		return err
+	}
+	icsCore, err := icscore.NewContract(ethcommon.HexToAddress(addresses.ICSCore), ethClient)
+	if err != nil {
+		return err
+	}
+	faucet, err := crypto.ToECDSA(ethcommon.FromHex(ethPrivateKey))
+	if err != nil {
+		return err
+	}
+	eth, err := ethereum.NewEthereum(context.Background(), ethereumRPC, nil, faucet)
+	if err != nil {
+		return err
+	}
+
+	// Connect to the Celestia prover
+	conn, err := grpc.Dial("localhost:50051", grpc.WithInsecure())
+	if err != nil {
+		return fmt.Errorf("failed to connect to prover: %w", err)
+	}
+	defer conn.Close()
+
+	fmt.Printf("Getting celestia prover info...\n")
+	proverClient := proverclient.NewProverClient(conn)
+	info, err := proverClient.Info(context.Background(), &proverclient.InfoRequest{})
+	if err != nil {
+		return fmt.Errorf("failed to get celestia prover info %w", err)
+	}
+	verifierKey := info.StateTransitionVerifierKey
+	fmt.Printf("Got celestia prover info. State transition verifier key %v", verifierKey)
+	trimmed := strings.TrimPrefix(verifierKey, "0x")
+	result, err := hex.DecodeString(trimmed)
+	if err != nil {
+		return fmt.Errorf("failed to decode verifier key %w", err)
+	}
+	// Convert the verifierKey byte slice into a [32]byte array
+	var VKey [32]byte
+	copy(VKey[:], result)
+	fmt.Printf("VKey %v\n", VKey)
+
+	request := &proverclient.ProveStateTransitionRequest{ClientId: addresses.ICS07Tendermint}
+	// Get state transition proof from Celestia prover with retry logic
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
+	defer cancel()
+	var resp *proverclient.ProveStateTransitionResponse
+	for retries := 0; retries < 3; retries++ {
+		resp, err = proverClient.ProveStateTransition(ctx, request)
+		if err == nil {
+			break
+		}
+		if ctx.Err() != nil {
+			return fmt.Errorf("context cancelled while getting state transition proof: %w", ctx.Err())
+		}
+		time.Sleep(time.Second * time.Duration(retries+1))
+	}
+	if err != nil {
+		return fmt.Errorf("failed to get state transition proof after retries: %w", err)
+	}
+	fmt.Printf("got resp %v\n", resp)
+	fmt.Printf("got resp.proof %v\n", hex.EncodeToString(resp.Proof))
+
+	msg := sp1ics07tendermint.IUpdateClientMsgsMsgUpdateClient{
+		Sp1Proof: sp1ics07tendermint.ISP1MsgsSP1Proof{
+			VKey:         VKey,
+			PublicValues: resp.PublicValues,
+			Proof:        resp.Proof,
+		},
+	}
+
+	arguments, err := getUpdateClientArguments()
+	if err != nil {
+		return err
+	}
+	fmt.Printf("Packing msg...\n")
+	encoded, err := arguments.Pack(msg)
+	if err != nil {
+		return fmt.Errorf("error packing msg %w", err)
+	}
+	fmt.Printf("Pcaked msg %v\n", encoded)
+
+	fmt.Printf("Invoking icsCore.UpdateClient...\n")
+	tx, err := icsCore.UpdateClient(getTransactOpts(faucet, eth), clientID, encoded)
+	if err != nil {
+		return err
+	}
+	fmt.Printf("icsCore.UpdateClient did not error\n")
+	receipt := getTxReciept(context.Background(), eth, tx.Hash())
+	if ethtypes.ReceiptStatusSuccessful != receipt.Status {
+		fmt.Printf("receipt %v\n", receipt)
+		fmt.Printf("receipt logs %v\n", receipt.Logs)
+		return fmt.Errorf("receipt status want %v, got %v", ethtypes.ReceiptStatusSuccessful, receipt.Status)
+	}
+	recvBlockNumber := receipt.BlockNumber.Uint64()
+	fmt.Printf("recvBlockNumber %v\n", recvBlockNumber)
+	return nil
+}
+
+func receivePacketOnEVM() error {
+	addresses, err := utils.ExtractDeployedContractAddresses()
+	if err != nil {
+		return err
+	}
+
+	sendPacket, err := createSendPacket()
+	if err != nil {
+		return err
+	}
+
+	packetCommitmentPath := ibchostv2.PacketCommitmentKey(sourceChannel, sequence)
+	fmt.Printf("packetCommitmentPath %v\n", packetCommitmentPath)
+
+	packet := ics26router.IICS26RouterMsgsPacket{
+		Sequence:         uint32(sendPacket.Sequence),
+		SourceChannel:    sendPacket.SourceChannel,
+		DestChannel:      sendPacket.DestinationChannel,
+		TimeoutTimestamp: sendPacket.TimeoutTimestamp,
+		Payloads: []ics26router.IICS26RouterMsgsPayload{
+			{
+				SourcePort: sendPacket.Payloads[0].SourcePort,
+				DestPort:   sendPacket.Payloads[0].DestinationPort,
+				Version:    transfertypes.V1,
+				Encoding:   transfertypes.EncodingABI,
+				Value:      sendPacket.Payloads[0].Value,
+			},
+		},
+	}
+
+	// TODO: replace this with query to celestia-prover after mock circuits
+	// are implemented.
+	membershipProof := []byte{}
+
+	// TODO: replace this with a real proof height.
+	proofHeight := ics26router.IICS02ClientMsgsHeight{
+		RevisionNumber: uint32(0),
+		RevisionHeight: uint32(10),
+	}
+	msg := ics26router.IICS26RouterMsgsMsgRecvPacket{
+		Packet:          packet,
+		ProofCommitment: membershipProof,
+		ProofHeight:     proofHeight,
+	}
+
+	ethClient, err := ethclient.Dial(ethereumRPC)
+	if err != nil {
+		return err
+	}
+	ics26Contract, err := ics26router.NewContract(ethcommon.HexToAddress(addresses.ICS26Router), ethClient)
+	if err != nil {
+		return err
+	}
+
+	faucet, err := crypto.ToECDSA(ethcommon.FromHex(ethPrivateKey))
+	if err != nil {
+		return err
+	}
+	eth, err := ethereum.NewEthereum(context.Background(), ethereumRPC, nil, faucet)
+	if err != nil {
+		return err
+	}
+	tx, err := ics26Contract.RecvPacket(getTransactOpts(faucet, eth), msg)
+	if err != nil {
+		return err
+	}
+
+	receipt := getTxReciept(context.Background(), eth, tx.Hash())
+	if ethtypes.ReceiptStatusSuccessful != receipt.Status {
+		return fmt.Errorf("receipt status want %v, got %v", ethtypes.ReceiptStatusSuccessful, receipt.Status)
+	}
+	recvBlockNumber := receipt.BlockNumber.Uint64()
+	fmt.Printf("recvBlockNumber %v\n", recvBlockNumber)
+	return nil
+}
+
+// TODO: refactor this to de-duplicate code from createMsgSendPacket
+func createSendPacket() (channeltypesv2.Packet, error) {
+	coin := sdktypes.NewCoin(denom, math.NewInt(100))
+	transferPayload := ics20lib.ICS20LibFungibleTokenPacketData{
+		Denom:    coin.Denom,
+		Amount:   coin.Amount.BigInt(),
+		Sender:   sender,
+		Receiver: receiver,
+		Memo:     "test transfer",
+	}
+	transferBz, err := ics20lib.EncodeFungibleTokenPacketData(transferPayload)
+	if err != nil {
+		return channeltypesv2.Packet{}, err
+	}
+	payload := channeltypesv2.Payload{
+		SourcePort:      transfertypes.PortID,
+		DestinationPort: transfertypes.PortID,
+		Version:         transfertypes.V1,
+		Encoding:        transfertypes.EncodingABI,
+		Value:           transferBz,
+	}
+
+	return channeltypesv2.Packet{
+		Sequence:           sequence,
+		SourceChannel:      ibctesting.FirstChannelID,
+		DestinationChannel: ibctesting.FirstClientID,
+		TimeoutTimestamp:   uint64(time.Now().Add(30 * time.Minute).Unix()),
+		Payloads:           []channeltypesv2.Payload{payload},
+	}, nil
+}
+
+func getTransactOpts(key *ecdsa.PrivateKey, chain ethereum.Ethereum) *bind.TransactOpts {
+	ethClient, err := ethclient.Dial(chain.RPC)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	fromAddress := crypto.PubkeyToAddress(key.PublicKey)
+	nonce, err := ethClient.PendingNonceAt(context.Background(), fromAddress)
+	if err != nil {
+		nonce = 0
+	}
+
+	gasPrice, err := ethClient.SuggestGasPrice(context.Background())
+	if err != nil {
+		panic(err)
+	}
+
+	txOpts, err := bind.NewKeyedTransactorWithChainID(key, chain.ChainID)
+	if err != nil {
+		log.Fatal(err)
+	}
+	txOpts.Nonce = big.NewInt(int64(nonce))
+	txOpts.GasPrice = gasPrice
+
+	// Set a specific gas limit
+	txOpts.GasLimit = 3000000 // Example gas limit; adjust as needed
+
+	return txOpts
+}
+
+func getTxReciept(ctx context.Context, chain ethereum.Ethereum, hash ethcommon.Hash) *ethtypes.Receipt {
+	ethClient, err := ethclient.Dial(chain.RPC)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	var receipt *ethtypes.Receipt
+	err = utils.WaitForCondition(time.Second*30, time.Second, func() (bool, error) {
+		receipt, err = ethClient.TransactionReceipt(ctx, hash)
+		if err != nil {
+			return false, nil
+		}
+
+		return receipt != nil, nil
+	})
+	if err != nil {
+		log.Fatalf("Failed to fetch receipt: %v", err)
+	}
+
+	// Log more details about the receipt
+	fmt.Printf("Transaction hash: %s\n", hash.Hex())
+	fmt.Printf("Block number: %d\n", receipt.BlockNumber.Uint64())
+	fmt.Printf("Gas used: %d\n", receipt.GasUsed)
+	fmt.Printf("Logs: %v\n", receipt.Logs)
+	if receipt.Status != ethtypes.ReceiptStatusSuccessful {
+		fmt.Println("Transaction failed. Inspect logs or contract.")
+	}
+
+	return receipt
+}
+
+func getUpdateClientArguments() (abi.Arguments, error) {
+	var updateClientABI = "[{\"type\":\"function\",\"name\":\"updateClient\",\"stateMutability\":\"pure\",\"inputs\":[{\"name\":\"o3\",\"type\":\"tuple\",\"internalType\":\"struct IUpdateClientMsgs.MsgUpdateClient\",\"components\":[{\"name\":\"sp1Proof\",\"type\":\"tuple\",\"internalType\":\"struct ISP1Msgs.SP1Proof\",\"components\":[{\"name\":\"vKey\",\"type\":\"bytes32\",\"internalType\":\"bytes32\"},{\"name\":\"publicValues\",\"type\":\"bytes\",\"internalType\":\"bytes\"},{\"name\":\"proof\",\"type\":\"bytes\",\"internalType\":\"bytes\"}]}]}],\"outputs\":[]}]"
+
+	parsed, err := abi.JSON(strings.NewReader(updateClientABI))
 	if err != nil {
 		return nil, err
 	}
 
-	queryClient := channeltypesv2.NewQueryClient(clientCtx)
-	request := channeltypesv2.QueryPacketCommitmentsRequest{ChannelId: channelID}
-	response, err := queryClient.PacketCommitments(context.Background(), &request)
-	if err != nil {
-		return nil, fmt.Errorf("failed to query packet commitments: %v", err)
-	}
-
-	fmt.Printf("Packet commitments: %v, packet height %v\n", response.GetCommitments(), response.GetHeight())
-	return response, nil
-}
-
-// GetStateTransitionProof returns a state transition proof from the Celestia
-// prover. The prover will query the Tendermint light client on the EVM roll-up
-// for it's last known height and generate a proof from that height all the way
-// up to the latest height on SimApp.
-func GetStateTransitionProof() (proof []byte, err error) {
-	conn, err := grpc.NewClient(celestiaProverEndpoint, grpc.WithTransportCredentials(insecure.NewCredentials()))
-	if err != nil {
-		log.Fatalf("Failed to connect to the prover service: %v", err)
-	}
-	defer conn.Close()
-
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second*60)
-	defer cancel()
-
-	client := proverclient.NewProverClient(conn)
-	request := &proverclient.ProveStateTransitionRequest{ClientId: ics07TMContractAddress}
-
-	resp, err := client.ProveStateTransition(ctx, request)
-	if err != nil {
-		return []byte{}, fmt.Errorf("failed to request state transition proof: %w", err)
-	}
-
-	fmt.Printf("Got state transition proof: %x, public values %v\n", resp.GetProof(), resp.GetPublicValues())
-	return resp.GetProof(), nil
-}
-
-// GetMembershipProof gets a membership proof that the packets in the input are
-// present in the state root at the input block height on SimApp.
-func GetMembershipProof(input *channeltypesv2.QueryPacketCommitmentsResponse) (proof []byte, err error) {
-	conn, err := grpc.NewClient(celestiaProverEndpoint, grpc.WithTransportCredentials(insecure.NewCredentials()))
-	if err != nil {
-		return []byte{}, fmt.Errorf("failed to connect to the prover service: %w", err)
-	}
-	defer conn.Close()
-
-	client := proverclient.NewProverClient(conn)
-	// Are packet commitments the correct data type to be proving here?
-	// TODO: investigate existing IBC relayer implementations.
-	request := &proverclient.ProveStateMembershipRequest{
-		Height:   int64(input.GetHeight().RevisionHeight),
-		KeyPaths: getKeyPaths(input.Commitments),
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second*60)
-	defer cancel()
-
-	response, err := client.ProveStateMembership(ctx, request)
-	if err != nil {
-		return []byte{}, fmt.Errorf("failed to request state membership proof: %w", err)
-	}
-
-	fmt.Printf("Got membership proof: %x, height %v\n", response.GetProof(), response.GetHeight())
-	return response.GetProof(), nil
-}
-
-// getKeyPaths returns a list of strings where each string is a Merkle path for
-// a leaf to the state root.
-func getKeyPaths(_ []*channeltypesv2.PacketState) []string {
-	// TODO: implement
-	return []string{}
+	return parsed.Methods["updateClient"].Inputs, nil
 }
